requirements.txt
anyio==4.9.0
babel==2.17.0
backrefs==5.9
beautifulsoup4==4.13.4
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
feedparser==6.0.11
ghp-import==2.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
Markdown==3.8.2
MarkupSafe==3.0.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.6.15
mkdocs-material-extensions==1.3.1
packaging==25.0
paginate==0.5.7
pathspec==0.12.1
platformdirs==4.3.8
Pygments==2.19.2
pymdown-extensions==10.16
python-dateutil==2.9.0.post0
PyYAML==6.0.2
pyyaml_env_tag==1.1
requests==2.32.4
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
typing_extensions==4.14.1
urllib3==2.5.0
watchdog==6.0.0
